Iteration 5. Create new Controller class and move all logic from Facade class
